Split into plan.rs, r2r.rs
